rename, remove size prop
